Move dependencies to fix CI
